Moved transpile switch to Lua 5.2
Closes #194
